Refactor to standalone modal component
