+++ public/pages/workflows/new_workflow/quick_configure_modal.tsx
@@ -1,0 +1,119 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import React, { useState } from 'react';
+import { useHistory } from 'react-router-dom';
+import {
+  EuiSmallButton,
+  EuiModal,
+  EuiModalHeader,
+  EuiModalHeaderTitle,
+  EuiModalBody,
+  EuiModalFooter,
+  EuiSmallButtonEmpty,
+  EuiCompressedFieldText,
+  EuiCompressedFormRow,
+} from '@elastic/eui';
+import { WORKFLOW_NAME_REGEXP, Workflow } from '../../../../common';
+import { APP_PATH } from '../../../utils';
+import { processWorkflowName } from './utils';
+import { createWorkflow, useAppDispatch } from '../../../store';
+import { constructUrlWithParams, getDataSourceId } from '../../../utils/utils';
+
+interface QuickConfigureModalProps {
+  workflow: Workflow;
+  onClose(): void;
+}
+
+export function QuickConfigureModal(props: QuickConfigureModalProps) {
+  const dispatch = useAppDispatch();
+  const dataSourceId = getDataSourceId();
+  const history = useHistory();
+
+  // workflow name state
+  const [workflowName, setWorkflowName] = useState<string>(
+    processWorkflowName(props.workflow.name)
+  );
+
+  // is creating state
+  const [isCreating, setIsCreating] = useState<boolean>(false);
+
+  // custom sanitization on workflow name
+  function isInvalid(name: string): boolean {
+    return (
+      name === '' ||
+      name.length > 100 ||
+      WORKFLOW_NAME_REGEXP.test(name) === false
+    );
+  }
+
+  return (
+    <EuiModal onClose={() => props.onClose()}>
+      <EuiModalHeader>
+        <EuiModalHeaderTitle>
+          <p>{`Set a unique name for your workflow`}</p>
+        </EuiModalHeaderTitle>
+      </EuiModalHeader>
+      <EuiModalBody>
+        <EuiCompressedFormRow
+          label={'Name'}
+          error={'Invalid name'}
+          isInvalid={isInvalid(workflowName)}
+        >
+          <EuiCompressedFieldText
+            placeholder={processWorkflowName(props.workflow.name)}
+            value={workflowName}
+            onChange={(e) => {
+              setWorkflowName(e.target.value);
+            }}
+          />
+        </EuiCompressedFormRow>
+      </EuiModalBody>
+      <EuiModalFooter>
+        <EuiSmallButtonEmpty onClick={() => props.onClose()}>
+          Cancel
+        </EuiSmallButtonEmpty>
+        <EuiSmallButton
+          disabled={isInvalid(workflowName) || isCreating}
+          isLoading={isCreating}
+          onClick={() => {
+            setIsCreating(true);
+            const workflowToCreate = {
+              ...props.workflow,
+              name: workflowName,
+            };
+            dispatch(
+              createWorkflow({
+                apiBody: workflowToCreate,
+                dataSourceId,
+              })
+            )
+              .unwrap()
+              .then((result) => {
+                setIsCreating(false);
+                const { workflow } = result;
+                history.replace(
+                  constructUrlWithParams(
+                    APP_PATH.WORKFLOWS,
+
+                    workflow.id,
+                    dataSourceId
+                  )
+                );
+              })
+              .catch((err: any) => {
+                setIsCreating(false);
+                console.error(err);
+              });
+          }}
+          fill={true}
+          color="primary"
+        >
+          Create
+        </EuiSmallButton>
+      </EuiModalFooter>
+    </EuiModal>
+  );
+}
