+++ public/pages/workflows/new_workflow/use_case.tsx
@@ -4,7 +4,6 @@
  */
 
 import React, { useState } from 'react';
-import { useHistory } from 'react-router-dom';
 import {
   EuiText,
   EuiFlexGroup,
@@ -13,118 +12,25 @@ import {
   EuiCard,
   EuiHorizontalRule,
   EuiSmallButton,
-  EuiModal,
-  EuiModalHeader,
-  EuiModalHeaderTitle,
-  EuiModalBody,
-  EuiModalFooter,
-  EuiSmallButtonEmpty,
-  EuiCompressedFieldText,
-  EuiCompressedFormRow,
 } from '@elastic/eui';
-import { WORKFLOW_NAME_REGEXP, Workflow } from '../../../../common';
-import { APP_PATH } from '../../../utils';
-import { processWorkflowName } from './utils';
-import { createWorkflow, useAppDispatch } from '../../../store';
-import { constructUrlWithParams, getDataSourceId } from '../../../utils/utils';
+import { Workflow } from '../../../../common';
+import { QuickConfigureModal } from './quick_configure_modal';
 
 interface UseCaseProps {
   workflow: Workflow;
 }
 
 export function UseCase(props: UseCaseProps) {
-  const dispatch = useAppDispatch();
-  const dataSourceId = getDataSourceId();
-  const history = useHistory();
-
   // name modal state
   const [isNameModalOpen, setIsNameModalOpen] = useState<boolean>(false);
-
-  // workflow name state
-  const [workflowName, setWorkflowName] = useState<string>(
-    processWorkflowName(props.workflow.name)
-  );
-
-  // is creating state
-  const [isCreating, setIsCreating] = useState<boolean>(false);
-
-  // custom sanitization on workflow name
-  function isInvalid(name: string): boolean {
-    return (
-      name === '' ||
-      name.length > 100 ||
-      WORKFLOW_NAME_REGEXP.test(name) === false
-    );
-  }
 
   return (
     <>
       {isNameModalOpen && (
-        <EuiModal onClose={() => setIsNameModalOpen(false)}>
-          <EuiModalHeader>
-            <EuiModalHeaderTitle>
-              <p>{`Set a unique name for your workflow`}</p>
-            </EuiModalHeaderTitle>
-          </EuiModalHeader>
-          <EuiModalBody>
-            <EuiCompressedFormRow
-              label={'Name'}
-              error={'Invalid name'}
-              isInvalid={isInvalid(workflowName)}
-            >
-              <EuiCompressedFieldText
-                placeholder={processWorkflowName(props.workflow.name)}
-                value={workflowName}
-                onChange={(e) => {
-                  setWorkflowName(e.target.value);
-                }}
-              />
-            </EuiCompressedFormRow>
-          </EuiModalBody>
-          <EuiModalFooter>
-            <EuiSmallButtonEmpty onClick={() => setIsNameModalOpen(false)}>
-              Cancel
-            </EuiSmallButtonEmpty>
-            <EuiSmallButton
-              disabled={isInvalid(workflowName) || isCreating}
-              isLoading={isCreating}
-              onClick={() => {
-                setIsCreating(true);
-                const workflowToCreate = {
-                  ...props.workflow,
-                  name: workflowName,
-                };
-                dispatch(
-                  createWorkflow({
-                    apiBody: workflowToCreate,
-                    dataSourceId,
-                  })
-                )
-                  .unwrap()
-                  .then((result) => {
-                    setIsCreating(false);
-                    const { workflow } = result;
-                    history.replace(
-                      constructUrlWithParams(
-                        APP_PATH.WORKFLOWS,
-
-                        workflow.id,
-                        dataSourceId
-                      )
-                    );
-                  })
-                  .catch((err: any) => {
-                    setIsCreating(false);
-                    console.error(err);
-                  });
-              }}
-              fill={true}
-              color="primary"
-            >
-              Create
-            </EuiSmallButton>
-          </EuiModalFooter>
-        </EuiModal>
+        <QuickConfigureModal
+          workflow={props.workflow}
+          onClose={() => setIsNameModalOpen(false)}
+        />
       )}
       <EuiCard
         title={
